tf2.0/requirements-gpu.txt
jupyter==1.0.0
numpy==1.14.5
pandas==0.23.4
scikit-learn==0.19.2
scipy==1.1.0
tensorflow-gpu==2.0.0-alpha0
requests
